Refactor proj definition schema + edit schema file names
