+++ src/snowcli/cli/project/schemas/relaxed_map.py
@@ -1,0 +1,44 @@
+from __future__ import annotations
+
+from strictyaml import (
+    Any,
+    Bool,
+    Decimal,
+    Int,
+    MapCombined,
+    Optional,
+    Str,
+)
+
+# TODO: use the util regexes to validate paths + globs
+FilePath = Str
+Glob = Str
+
+
+class RelaxedMap(MapCombined):
+    """
+    A version of a Map that allows any number of unknown key/value pairs.
+    """
+
+    def __init__(self, map_validator):
+        super().__init__(
+            map_validator,
+            Str(),
+            # moves through value validators left-to-right until one matches
+            Bool() | Decimal() | Int() | Any(),
+        )
+
+    def as_fully_optional(self) -> RelaxedMap:
+        """
+        Returns a copy of this schema with all its keys optional, recursing into other
+        RelaxedMaps we find inside the schema. For existing optional keys, we strip out
+        the default value and ensure we don't create any new keys.
+        """
+        validator = {}
+        for key, value in self._validator_dict.items():
+            validator[Optional(key)] = (
+                value
+                if not isinstance(value, RelaxedMap)
+                else value.as_fully_optional()
+            )
+        return RelaxedMap(validator)
